Refactor NoLight into modular package
